kcs-logger
***Players need to open the in-game Interface menu, navigate to the Game tab, and under the "Combat" section, manually check the option for "Enable Advanced Combat Logging." This manual step is intentional to ensure that players are aware and in control of the information being logged.***

 

kcs-logger, short for kc8pnd's logger:


When this addon is enabled in World of Warcraft, it will monitor the player's location and dynamically control combat logging. Upon entering specific zones (e.g., "Blackfathom Deeps," "Gnomeregan," "Zul'Gurub"), the addon will enable combat logging, allowing the game to record combat-related events. A message will be displayed in the default chat frame indicating that combat logging is enabled for the current zone.

Conversely, when the player exits the listed zones, the addon will disable combat logging and display a message confirming that combat logging is disabled. Essentially, the addon automates the process of enabling and disabling combat logging based on the player's location in the game world. This can be useful for selectively recording combat data in specific zones without the need for manual intervention by the player.